# rgw_sal_motr: [CORTX-29664] optimise head-object s3api handling