@andriytk 2022-03-31
Sorry, I failed to get the problem from the "Problem Description". Please elaborate.