@pranavPawar07 2022-04-04
> @pranavPawar07 , what about multipart uploaded objects? Are we trying to open the mobj there as well?

@sjain09 , yes