@pranavPawar07 2022-04-04
> Sorry, I failed to get the problem from the "Problem Description". Please elaborate.

@andriytk edited the problem description. Please check