@andriytk 2022-04-04
> > Sorry, I failed to get the problem from the "Problem Description". Please elaborate.
> 
> @andriytk edited the problem description. Please check

Please, rename the title of this PR and the commit subject line to something like: "rgw_sal_motr: [CORTX-29664] optimise head-object s3api handling". Refer to 3.d section at https://seagate-systems.atlassian.net/wiki/spaces/PUB/pages/939720719/Creating+RGW+Pull+Requests+-+Guideline#Two-key-questions for more info on this.