@pranavPawar07 2022-04-04
> > ```
> > * also, we were not closing the object which might result in a memory leak.
> > ```
> 
> This is interesting... Is it relevant only for head-object s3api handling?

Yes.