@andriytk 2022-04-04
Please, squash all the intermediate commits into a single, final one.

Also, add the problem description and the solution into the commit msg.